Enable virtual-host-style requests for buckets
By default, MinIO supports path-style requests that are of the format http://mydomain.com/bucket/object.
This PR adds an optional MINIO_DOMAIN environment variable that is used to enable virtual-host-style requests.
If the request Host header matches with (.+).mydomain.com then the matched pattern $1 is used as bucket and the path is used as object.